requirements.txt
aioice==0.7.5
aiortc==1.2.1
av==8.0.3
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
dnspython==2.1.0
Faker==8.8.2
Flask==2.0.1
google-crc32c==1.1.2
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
netifaces==0.11.0
playsound==1.2.2
pycparser==2.20
pyee==8.1.0
pylibsrtp==0.6.8
python-dateutil==2.8.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
text-unidecode==1.3
undetected-chromedriver==3.0.2
urllib3==1.26.5
websockets==9.1
Werkzeug==2.0.1
